pyproject.toml
[project]
name = "stonks-lake"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Michal", email = "mcpiatkowski@gmail.com"},
]
dependencies = [
    "aws-cdk-lib>=2.158.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
